Removal of quarterly vs semiannual coupon flag
Flag removed from "D_Setup" collection of settings and deleted as argument in functions appearing in Notebook. Argument is set to default of 2 (semiannual coupons) in relevant functions of bin library: Make_DbtPaths() and Performance(). Notice that now it is no longer possible to closely replicate Belton et al, which had quarterly coupons.